test(activerecord): move describeIfSqlite to support so both test trees share it
