+++ packages/activerecord/src/adapters/sqlite3/explain.test.ts
@@ -3,7 +3,7 @@
  */
 import { it, expect } from "vitest";
 import "../../index.js";
-import { describeIfSqlite } from "./test-helper.js";
+import { describeIfSqlite } from "../../support/describe-if-sqlite.js";
 import { fixtures } from "../../test-fixtures.js";
 import { Author } from "../../test-helpers/models/author.js";
 // Opt into the canonical-model autoload index so `Author.has_many :posts`
