+++ packages/activerecord/src/adapters/sqlite3/statement-pool.trails.test.ts
@@ -1,5 +1,5 @@
 import { it, expect, afterEach } from "vitest";
-import { describeIfSqlite } from "./test-helper.js";
+import { describeIfSqlite } from "../../support/describe-if-sqlite.js";
 import { AbstractSQLite3Adapter } from "../../connection-adapters/sqlite3-adapter.js";
 import { BetterSQLite3Adapter } from "../../connection-adapters/better-sqlite3-adapter.js";
