+++ packages/activerecord/src/adapters/sqlite3/test-helper.ts
@@ -2,22 +2,9 @@
  * Shared helpers for `adapters/sqlite3/*.test.ts`. Keep this file tiny —
  * anything beyond cross-test glue belongs in the source tree, not here.
  */
-import { describe, expect } from "vitest";
+import { expect } from "vitest";
 import { Notifications, squish } from "@blazetrails/activesupport";
 import type { NotificationEvent } from "@blazetrails/activesupport";
-import { isSqliteRun } from "../../support/sqlite-template.js";
-
-/**
- * Scope a suite to the SQLite backend, mirroring Rails'
- * `ActiveRecord::SQLite3TestCase`. The handler connection is swapped to
- * PG/MySQL in the cross-backend CI matrix, so suites whose assertions are
- * SQLite-specific — `EXPLAIN QUERY PLAN` plan shape, `"`-quoted identifiers —
- * must skip there rather than run against a backend Rails never points them at.
- * Reuses the canonical {@link isSqliteRun} predicate so the "what counts as a
- * SQLite run" rule has one source of truth. Counterpart to `describeIfPg` /
- * `describeIfMysql`.
- */
-export const describeIfSqlite = isSqliteRun() ? describe : (describe.skip as typeof describe);
 
 /**
  * Subscribe to `sql.active_record` for the duration of `fn`, then assert the
