+++ packages/activerecord/src/adapters/sqlite3/virtual-column.trails.test.ts
@@ -1,6 +1,6 @@
 import { it, expect, beforeEach, afterEach } from "vitest";
 import "../../index.js";
-import { describeIfSqlite } from "./test-helper.js";
+import { describeIfSqlite } from "../../support/describe-if-sqlite.js";
 import { Base } from "../../base.js";
 import { fixtures } from "../../test-fixtures.js";
 import type { AbstractSQLite3Adapter } from "../../connection-adapters/sqlite3-adapter.js";
