+++ packages/activerecord/src/connection-adapters/sqlite3-adapter.query-transformers.test.ts
@@ -2,7 +2,7 @@ import { it, expect, beforeEach, afterEach } from "vitest";
 import { Notifications } from "@blazetrails/activesupport";
 import { Base } from "../base.js";
 import { fixtures } from "../test-fixtures.js";
-import { describeIfSqlite } from "../adapters/sqlite3/test-helper.js";
+import { describeIfSqlite } from "../support/describe-if-sqlite.js";
 import type { AbstractSQLite3Adapter } from "./sqlite3-adapter.js";
 import { queryTransformers, type QueryTransformer } from "../query-transformers.js";
