+++ packages/activerecord/src/connection-adapters/sqlite3-copy-table.test.ts
@@ -1,7 +1,7 @@
 import { it, expect, beforeEach, afterEach } from "vitest";
 import { Base } from "../base.js";
 import { fixtures } from "../test-fixtures.js";
-import { describeIfSqlite } from "../adapters/sqlite3/test-helper.js";
+import { describeIfSqlite } from "../support/describe-if-sqlite.js";
 import type { AbstractSQLite3Adapter } from "./sqlite3-adapter.js";
 
 fixtures([]);
