+++ packages/activerecord/src/defaults.test.ts
@@ -19,7 +19,7 @@ import {
   MYSQL_TEST_URL,
 } from "./adapters/abstract-mysql-adapter/test-helper.js";
 import { describeIfPg } from "./adapters/postgresql/test-helper.js";
-import { describeIfSqlite } from "./adapters/sqlite3/test-helper.js";
+import { describeIfSqlite } from "./support/describe-if-sqlite.js";
 import { describeIfSupports, itIfSupports } from "./support/supports.js";
 import { fixtures } from "./test-fixtures.js";
 import { Entrant } from "./test-helpers/models/entrant.js";
